conf: CROSSFADE must be an int

+++ teamplayer/conf.py
@@ -34,7 +34,7 @@ class TeamPlayerSettings:
     SCROBBLER_PASSWORD: str = ""
     LASTFM_APIKEY: str = "2d5952c801e074e3251bafb77f54e680"
     LASTFM_APISECRET: str = "23dd84f772d374d7f8230d74afc8d269"
-    CROSSFADE: float = 0.0
+    CROSSFADE: int = 0
     SHAKE_THINGS_UP: int = 0
     SHAKE_THINGS_UP_FILTER: str = '{"length__lt": 300, "length__gt": 0}'
     ALWAYS_SHAKE_THINGS_UP: bool = False
